+++ src/ContaoFigcaptionBundle.php
@@ -1,5 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * Dieses Bundle ersetzt in Bildunterschriften einen von Trennzeichen
+ * umklammerten Text, damit sich z. B. der Fotograf gesondert auszeichnen lässt.
+ *
+ * @author    Frank Binding
+ * @license   LGPL-3.0-or-later
+ */
+
 namespace Schachbulle\ContaoFigcaptionBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;